feat: add iOS/Android mobile cross-compilation support
- CMakeLists.txt: platform detection (TASK_GRAPH_IOS/ANDROID/MOBILE),
  STATIC lib for mobile, guard tests/examples/ctest, OpenCV mobile paths
- src/plugin.cpp: iOS dlopen guard (TASK_GRAPH_NO_DLOPEN stubs + early return)
- src/logger.cpp: Android gettid() branch, skip pthread_getname_np (bionic)
- src/profiler.cpp: cast tid to long long (Android int64_t=long ambiguity)
- Submodule CMakeLists: STATIC for mobile (EMSCRIPTEN OR TASK_GRAPH_MOBILE)
- scripts/build_ios.sh: iOS arm64 device+sim, libtool merge, xcframework
- scripts/build_android.sh: NDK cross-compile, llvm-ar MRI merge
- scripts/build_opencv_{ios,android}.sh: OpenCV static lib cross-build

+++ src/profiler.cpp
@@ -321,7 +321,7 @@ std::string ProfileCollector::to_trace_string(bool pretty) const {
     };
 
     for (const auto& ts : task_stats) {
-        int64_t tid = tid_of(ts.task_id);
+        long long tid = static_cast<long long>(tid_of(ts.task_id));
 
         // 等待段（READY → STARTED）：浅色，类别 "wait"
         if (ts.has_ready && ts.has_start && ts.start_time > ts.ready_time) {
